Substantial changes to Primer.
Primer now takes a GrinClient at instantiation. It uses the
client to query GRIN for the list of converted books and
compares that list with those that have already been downloaded
to the processing_bucket. It then creates tickets for those
that have not already been downloaded and puts them in the
"to_bucket" for the pipeline.